rename "client" file
